+++ worker/src/commom_api.ts
@@ -15,6 +15,11 @@ api.get('/open_api/settings', async (c) => {
         const auth = c.req.raw.headers.get("x-custom-auth");
         needAuth = !auth || !passwords.includes(auth);
     }
+    const smtpImapProxyConfig = utils.getJsonObjectValue<SmtpImapProxyConfig>(
+        c.env.SMTP_IMAP_PROXY_CONFIG
+    ) || {};
+    const smtpProxyConfig = smtpImapProxyConfig.smtp || {};
+    const imapProxyConfig = smtpImapProxyConfig.imap || {};
 
     return c.json({
         "title": c.env.TITLE,
@@ -45,6 +50,19 @@ api.get('/open_api/settings', async (c) => {
         "showGithub": !utils.getBooleanValue(c.env.DISABLE_SHOW_GITHUB),
         "disableAdminPasswordCheck": utils.getBooleanValue(c.env.DISABLE_ADMIN_PASSWORD_CHECK),
         "enableAddressPassword": utils.getBooleanValue(c.env.ENABLE_ADDRESS_PASSWORD),
+        "enableAgentEmailInfo": utils.getBooleanValue(c.env.ENABLE_AGENT_EMAIL_INFO),
+        "smtpImapProxyConfig": {
+            "smtp": {
+                "host": utils.getStringValue(smtpProxyConfig.host),
+                "port": utils.getIntValue(smtpProxyConfig.port, 8025),
+                "starttls": utils.getBooleanValue(smtpProxyConfig.starttls),
+            },
+            "imap": {
+                "host": utils.getStringValue(imapProxyConfig.host),
+                "port": utils.getIntValue(imapProxyConfig.port, 11143),
+                "starttls": utils.getBooleanValue(imapProxyConfig.starttls),
+            },
+        },
         "statusUrl": utils.getStringValue(c.env.STATUS_URL),
         "enableGlobalTurnstileCheck": utils.isGlobalTurnstileEnabled(c)
     });